# go-web/docs/swagger.yaml
basePath: /api/v1
definitions:
  handler.userRequest:
    properties:
      edad:
        type: integer
      estaActive:
        type: boolean
      nombre:
        type: string
    required:
    - edad
    - nombre
    type: object
  handler.userRequestPatch:
    properties:
      edad:
        type: integer
      nombre:
        type: string
    required:
    - edad
    - nombre
    type: object
  web.Response:
    properties:
      code:
        type: string
      data: {}
      error:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support Dionys Seidel
    url: http://www.swagger.io/support
  description: This API Handle MELI Users. This is a simple API development conducted
    by Digital House's team.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://developers.mercadolibre.com.ar/es_ar/terminos-y-condiciones
  title: Bootcamp Go Wave 6 - API
  version: "1.0"
paths:
  /users:
    get:
      consumes:
      - application/json
      description: get users
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List users
          schema:
            $ref: '#/definitions/web.Response'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/web.Response'
        "404":
          description: not found users
          schema:
            $ref: '#/definitions/web.Response'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/web.Response'
      summary: Show list users
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: store users
      parameters:
      - description: token requeridx
        in: header
        name: token
        required: true
        type: string
      - description: User to store
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handler.userRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/web.Response'
      summary: Store user
      tags:
      - Users
  /users/{id}:
    delete:
      parameters:
      - description: user ID
        in: path
        name: id
        required: true
        type: integer
      - description: token
        in: header
        name: token
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.Response'
      summary: Delete user
      tags:
      - Users
    put:
      consumes:
      - application/json
      parameters:
      - description: ID user
        in: path
        name: id
        required: true
        type: integer
      - description: token
        in: header
        name: token
        type: string
      - description: User to update
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handler.userRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.Response'
      summary: Update user
      tags:
      - Users
swagger: "2.0"